refactor: extract DbgpXml from DebugServer
Move DBGp XML sanitization into a focused final class so pure logic is
unit-tested without reflection, and DebugServer only delegates.
