+++ tests/Unit/DebugServerXmlSanitizerTest.php
@@ -4,9 +4,8 @@ declare(strict_types=1);
 
 namespace Koriym\XdebugMcp\Tests\Unit;
 
-use Koriym\XdebugMcp\DebugServer;
+use Koriym\XdebugMcp\Dbgp\DbgpXml;
 use PHPUnit\Framework\TestCase;
-use ReflectionMethod;
 
 use function libxml_clear_errors;
 use function libxml_use_internal_errors;
@@ -14,13 +13,6 @@ use function simplexml_load_string;
 
 class DebugServerXmlSanitizerTest extends TestCase
 {
-    private ReflectionMethod $sanitize;
-
-    protected function setUp(): void
-    {
-        $this->sanitize = new ReflectionMethod(DebugServer::class, 'sanitizeDbgpXml');
-    }
-
     public function testStripsNullByteFromAnonymousClassname(): void
     {
         $payload = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
@@ -28,7 +20,7 @@ class DebugServerXmlSanitizerTest extends TestCase
             . "\x00"
             . '/path/to/file.php:42"/>';
 
-        $clean = $this->sanitize->invoke(null, $payload);
+        $clean = DbgpXml::sanitize($payload);
 
         $this->assertStringNotContainsString("\x00", $clean);
 
@@ -44,19 +36,19 @@ class DebugServerXmlSanitizerTest extends TestCase
     public function testPreservesTabNewlineAndCarriageReturn(): void
     {
         $input = "a\tb\nc\rd";
-        $this->assertSame($input, $this->sanitize->invoke(null, $input));
+        $this->assertSame($input, DbgpXml::sanitize($input));
     }
 
     public function testPreservesUtf8MultibyteSequences(): void
     {
         $input = 'こんにちは';
-        $this->assertSame($input, $this->sanitize->invoke(null, $input));
+        $this->assertSame($input, DbgpXml::sanitize($input));
     }
 
     public function testStripsAdditionalC0Controls(): void
     {
         $input = "begin\x01\x05\x0B\x0C\x0E\x1F\x7Fend";
-        $this->assertSame('beginend', $this->sanitize->invoke(null, $input));
+        $this->assertSame('beginend', DbgpXml::sanitize($input));
     }
 
     public function testStripsInvalidNumericCharacterReferences(): void
@@ -66,7 +58,7 @@ class DebugServerXmlSanitizerTest extends TestCase
             . '&#0;&#x0;&#x0B;'
             . '/path/to/file.php:42" valid="&#9;&#10;&#13;&#x20;"/>';
 
-        $clean = $this->sanitize->invoke(null, $payload);
+        $clean = DbgpXml::sanitize($payload);
 
         $this->assertStringNotContainsString('&#0;', $clean);
         $this->assertStringNotContainsString('&#x0;', $clean);
@@ -87,7 +79,7 @@ class DebugServerXmlSanitizerTest extends TestCase
         $payload = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
             . '<response invalid="&#11;&#xD800;&#xDFFF;&#xFFFE;&#xFFFF;&#x110000;" valid="&#x10000;"/>';
 
-        $clean = $this->sanitize->invoke(null, $payload);
+        $clean = DbgpXml::sanitize($payload);
 
         $this->assertStringNotContainsString('&#11;', $clean);
         $this->assertStringNotContainsString('&#xD800;', $clean);
@@ -108,6 +100,19 @@ class DebugServerXmlSanitizerTest extends TestCase
     public function testStripsOverlongNumericCharacterReferences(): void
     {
         $input = 'begin&#xFFFFFFFFFFFF;&#99999999;end';
-        $this->assertSame('beginend', $this->sanitize->invoke(null, $input));
+        $this->assertSame('beginend', DbgpXml::sanitize($input));
+    }
+
+    public function testIsXmlCharacterAcceptsValidCodepoints(): void
+    {
+        $this->assertTrue(DbgpXml::isXmlCharacter(0x09));
+        $this->assertTrue(DbgpXml::isXmlCharacter(0x0A));
+        $this->assertTrue(DbgpXml::isXmlCharacter(0x0D));
+        $this->assertTrue(DbgpXml::isXmlCharacter(0x20));
+        $this->assertTrue(DbgpXml::isXmlCharacter(0x10000));
+        $this->assertFalse(DbgpXml::isXmlCharacter(0x00));
+        $this->assertFalse(DbgpXml::isXmlCharacter(0x0B));
+        $this->assertFalse(DbgpXml::isXmlCharacter(0xD800));
+        $this->assertFalse(DbgpXml::isXmlCharacter(0x110000));
     }
 }
